requirement.txt
Flask==3.0.0
google-generativeai==0.5.2
gunicorn==21.2.0